fix: move PeanutDoesntStoreAnyPersonalInformation out to a separate component
